Move irda_write to ISR and scroll end screen
Now publishing the IrDA packets from the new, second alarm
interrupt, freeing up the main loop to scroll the text on the end
screen.

+++ firmware/include/tetrominos.h
@@ -1,8 +1,10 @@
 #ifndef TETROMINOS_INCLUDED
 #define TETROMINOS_INCLUDED
 
+#include <stdint.h>
+
 typedef struct {
-    int x, y;
+    int8_t x, y;
 } vertex_t;
 
 typedef struct {
